query2_xnet_package/requirements.txt
wordcloud
numpy
six
nltk
python-igraph
pandas
matplotlib
pillow